third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/upload-onloadend-event-after-load.html

<!doctype html>
<html>
<head>
<title>Test case for bug 40952</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body>
<p> Test case for <a href="https://bugs.webkit.org/show_bug.cgi?id=40952"> bug 40952</a>: Onloadend event is not supported in XMLHttpRequest</p>
<p> Verify that a loadend ProgressEvent is dispatched after the load ProgressEvent when an async upload request completes normally.</p>
<div id="log"></div>
<script type="text/javascript">

var xhr;
var payload = "data";
var expectedLength = payload.length;
var results = "";
var expected = " loadstart load loadend";

function verifyProgressEvent(context, e, expected)
{
    assert_true(e.lengthComputable);
    assert_equals(e.loaded, expected, "Expected 'loaded' value for '" + context + "' event.");
    assert_equals(e.total, expected, "Expected 'total' value for '" + context + "' event.");
}

function logProgressEvent(e) {
    results += " " + e.type;
}

function logUnexpectedProgressEvent(e) {
    results += " [unexpected ProgressEvent: " + e.type + "]";
    completeTest();
}

function loadendHandler(e)
{
    logProgressEvent(e);
    assert_true(e instanceof ProgressEvent);
    verifyProgressEvent("onloadend", e, expectedLength);
    completeTest();
}

function completeTest()
{
    assert_equals(results, expected, "Expected load event sequence");
    testUploadOnloadEndEvent.done();
}

var testUploadOnloadEndEvent = async_test("Check that upload 'loadend' events are delivered and have expected values.");
testUploadOnloadEndEvent.step(function () {
    xhr = new XMLHttpRequest();
    xhr.upload.onloadstart = testUploadOnloadEndEvent.step_func(logProgressEvent);
    xhr.upload.onabort = testUploadOnloadEndEvent.step_func(logUnexpectedProgressEvent);
    xhr.upload.onerror = testUploadOnloadEndEvent.step_func(logUnexpectedProgressEvent);
    xhr.upload.onload = testUploadOnloadEndEvent.step_func(logProgressEvent);
    xhr.upload.onloadend = testUploadOnloadEndEvent.step_func(loadendHandler);
    xhr.open("POST", "resources/post-echo.php", true);
    xhr.send(payload);
});
</script>
</body>
</html>
